create components for dragon lair and syrups for performance
